pyproject.toml
# Project metadata
[tool.poetry]
name = "get-started"
version = "0.0.1"
description = "A simple example to get started with the restack-ai SDK"
authors = [
    "Restack Team <service@restack.io>",
]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
restack-ai = "^0.0.21"
streamlit = "^1.39.0"
pypdf = "^5.1.0"
vapi-server-sdk = "^0.1.0"
phonenumbers = "^8.13.48"
filelock = "^3.16.1"
weave = "^0.51.18"

[tool.poetry.dev-dependencies]
pytest = "^6.2"  # Optional: Add if you want to include tests in your example

# Build system configuration
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# CLI command configuration
[tool.poetry.scripts]
services = "src.services:run_services"
schedule = "schedule_workflow:main"
flask = "app:app"